Replace EventFdMap/mutex with self-contained EventFd struct
Introduce co::EventFd, a portable triggerable file descriptor that owns a
poll-side fd and a trigger-side fd. On Linux it wraps an eventfd and on
macOS a kqueue (both with poll_fd == trigger_fd); on other systems it
wraps a non-blocking pipe pair, removing the need for the global
EventFdMap and its mutex. Modeled after toolbelt::TriggerFd but
duplicated here to avoid a circular dependency on cpp_toolbelt.

NewEventFd, CloseEventFd, TriggerEvent, ClearEvent, and MakeAbortEvent
are now member functions/factories on EventFd. All places that used to
hold a raw event-fd int (Coroutine::abort_fd_, the per-coroutine event
fd, and the scheduler's interrupt_fd_/co_interrupt_fd_) now hold an
EventFd, which lets GetInterruptFd() collapse to a single
implementation. Also fix a pre-existing leak where the scheduler's
co_interrupt_fd_ was never closed.

Bump the module version to 3.3.0 and silence a -Wsign-compare warning
in http_client/main.cc.

Made-with: Cursor

+++ http_client/main.cc
@@ -134,8 +134,8 @@ static size_t ReadContents(co::Coroutine *c, int fd, std::string &buffer,
     if (i < buffer.size()) {
       // Data remaining in buffer
       size_t nbytes = buffer.size() - i;
-      if (nbytes > length) {
-        nbytes = length;
+      if (nbytes > static_cast<size_t>(length)) {
+        nbytes = static_cast<size_t>(length);
       }
       if (write_to_output) {
         fwrite(&buffer[i], 1, nbytes, stdout);
